+++ tests/submit-request-config-v200.test.js
@@ -12,7 +12,7 @@ const OUTING_HEADERS = [
   "require_leave_date", "require_return_date", "require_return_time", "require_guardian_phone",
   "require_guardian_relation", "require_emergency_reason", "require_purpose", "require_location",
   "require_vehicle", "require_warden_approval", "require_selfie", "config_version", "created_at",
-  "created_by", "updated_at", "updated_by"
+  "created_by", "updated_at", "updated_by", "departure_allowed_days", "earliest_departure_time"
 ];
 
 class FakeSheet {
@@ -132,6 +132,8 @@ function completeConfig(overrides = {}) {
     allowed_days: "ISNIN",
     application_open_time: "",
     application_close_time: "",
+    departure_allowed_days: "",
+    earliest_departure_time: "",
     fixed_return_time: "",
     same_day_only: false,
     require_leave_date: false,